Move EditPersonDescriptor to a general descriptor class for extensibility
